requirements.txt
requests~=2.23.0
tqdm~=4.29.1
pandas~=1.0.4
biopython>=1.77